requirements.txt
comtypes==1.1.8
MouseInfo==0.1.3
numpy==1.20.0
opencv-python==4.5.1.48
Pillow==8.1.0
psutil==5.8.0
PyAutoGUI==0.9.52
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.7.2
pyperclip==1.8.1
PyRect==0.1.4
PyScreeze==0.1.26
PyTweening==1.0.3
pywin32==300
pywinauto==0.6.8
pyWinhook==1.6.2
six==1.15.0
